requirements.txt
aiogram==3.7.0
mysql-connector-python==8.0.33
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
deep-translator==1.11.4
aiohttp==3.9.5
APScheduler==3.10.4 
openai==1.12.0 